Task 11 - Guvi Login Functionality
I have executed the code in headless browser, if need to check in normal browser please uncomment the lines and execute the code. 
Attached the html report generated along with the code for reference.
Have checked the "is_displayed" & "is_enabled" options for email and password fields.
For failure: In email field, password is entered and in password field email is entered intentionally and upon login, user does not land on the dashboard which is been checked.